118: Allow editing of one-time payments 118: Do some rewriting on how to handle counting of one time-payments in regards to subscribing to events. This should prevent users from subscribing to more events than the remaining amount of one-time payments

+++ api/src/main/java/io/oxalate/backend/api/AbstractPayment.java
@@ -26,5 +26,5 @@ public abstract class AbstractPayment {
 
     @Schema(description = "If type is ONE_TIME then this has to be positive integer", example = "4", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     @JsonProperty("paymentCount")
-    private Long paymentCount;
+    private Integer paymentCount;
 }
